# documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "description": "",
        "title": "Road Protect API",
        "version": "1.4.1"
    },
    "paths": {
...
        "/api/v1/query/account/spreadsheet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/AdvancedTableColumn"
                            },
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "type": "object",
                            "title":"Buffer"
                        }
                    },
                    "400": {
                        "description": "E112: No data to export to spreadsheet"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Invalid query parameters"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Accounts"
                ],
                "operationId": "AccountQueryController_getManyAsSpreadsheet",
                "summary": "Query and filter accounts as a spreadsheet"
            }
        },
...
    },
    "definitions": {
...
        "AdvancedTableColumn": {
            "title":"AdvancedTableColumn",
            "properties": {
                "isDisplaying": {
                    "default": true,
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "spreadsheetFormat": {
                    "default": "default",
                    "enum": [
                        "currency",
                        "date",
                        "number",
                        "default"
                    ],
                    "type": "string"
                },
                "spreadsheetKey": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            },
            "required": [
                "key",
                "title",
                "isDisplaying",
                "spreadsheetKey",
                "spreadsheetFormat"
            ],
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
...
}